+++ nautilus_trader/backtest/engine.pyx
@@ -14,6 +14,7 @@
 # -------------------------------------------------------------------------------------------------
 
 import socket
+from decimal import Decimal
 from typing import List, Optional, Union
 
 import pandas as pd
@@ -23,7 +24,6 @@ import pytz
 from cpython.datetime cimport datetime
 from libc.stdint cimport int64_t
 
-from nautilus_trader.analysis.performance cimport PerformanceAnalyzer
 from nautilus_trader.backtest.data_client cimport BacktestDataClient
 from nautilus_trader.backtest.data_client cimport BacktestMarketDataClient
 from nautilus_trader.backtest.exchange cimport SimulatedExchange
@@ -51,7 +51,7 @@ from nautilus_trader.execution.engine cimport ExecutionEngine
 from nautilus_trader.infrastructure.cache cimport RedisCacheDatabase
 from nautilus_trader.model.c_enums.account_type cimport AccountType
 from nautilus_trader.model.c_enums.aggregation_source cimport AggregationSource
-from nautilus_trader.model.c_enums.book_level cimport BookLevel
+from nautilus_trader.model.c_enums.book_type cimport BookType
 from nautilus_trader.model.c_enums.oms_type cimport OMSType
 from nautilus_trader.model.c_enums.venue_type cimport VenueType
 from nautilus_trader.model.data.bar cimport Bar
@@ -69,6 +69,7 @@ from nautilus_trader.risk.engine cimport RiskEngine
 from nautilus_trader.serialization.msgpack.serializer cimport MsgPackSerializer
 from nautilus_trader.trading.strategy cimport TradingStrategy
 
+from nautilus_trader.analysis.performance import PerformanceAnalyzer
 from nautilus_trader.cache.cache import CacheConfig
 from nautilus_trader.data.engine import DataEngineConfig
 from nautilus_trader.execution.engine import ExecEngineConfig
@@ -476,10 +477,13 @@ cdef class BacktestEngine:
         AccountType account_type,
         Currency base_currency,
         list starting_balances,
+        default_leverage=None,
+        dict leverages=None,
         bint is_frozen_account=False,
         list modules=None,
         FillModel fill_model=None,
-        BookLevel order_book_level=BookLevel.L1,
+        BookType book_type=BookType.L1_TBBO,
+        fill_limit_at_price=False,
     ) -> None:
         """
         Add a `SimulatedExchange` with the given parameters to the backtest engine.
@@ -499,14 +503,20 @@ cdef class BacktestEngine:
             The account base currency for the client. Use ``None`` for multi-currency accounts.
         starting_balances : list[Money]
             The starting account balances (specify one for a single asset account).
+        default_leverage : Decimal
+            The account default leverage (for margin accounts).
+        leverages : Dict[InstrumentId, Decimal]
+            The instrument specific leverage configuration (for margin accounts).
         is_frozen_account : bool
             If the account for this exchange is frozen (balances will not change).
         modules : list[SimulationModule, optional
             The simulation modules to load into the exchange.
         fill_model : FillModel, optional
             The fill model for the exchange (if None then no probabilistic fills).
-        order_book_level : BookLevel
-            The default order book level for fill modelling.
+        book_type : BookType
+            The default order book type for fill modelling.
+        fill_limit_at_price : bool
+            If limit orders should be filled at their original price only (overrides slippage).
 
         Raises
         ------
@@ -532,14 +542,17 @@ cdef class BacktestEngine:
             account_type=account_type,
             base_currency=base_currency,
             starting_balances=starting_balances,
+            default_leverage=default_leverage or Decimal(10),
+            leverages=leverages or {},
             is_frozen_account=is_frozen_account,
             instruments=self._cache.instruments(venue),
             modules=modules,
             cache=self._cache,
             fill_model=fill_model,
-            exchange_order_book_level=order_book_level,
+            book_type=book_type,
             clock=self._test_clock,
             logger=self._test_logger,
+            fill_limit_at_price=fill_limit_at_price,
         )
 
         self._exchanges[venue] = exchange
@@ -703,6 +716,7 @@ cdef class BacktestEngine:
         self._test_clock.set_time(start_ns)
         self._test_logger.change_clock_c(self._test_clock)
 
+        cdef SimulatedExchange exchange
         if not streaming:
             for exchange in self._exchanges.values():
                 exchange.initialize_account()
@@ -738,18 +752,26 @@ cdef class BacktestEngine:
         cdef Data data = self._next()
         while data is not None:
             self._advance_time(data.ts_init)
+            self._data_engine.process(data)
             if isinstance(data, OrderBookData):
                 self._exchanges[data.instrument_id.venue].process_order_book(data)
             elif isinstance(data, Tick):
                 self._exchanges[data.instrument_id.venue].process_tick(data)
-            self._data_engine.process(data)
-            self._process_modules(data.ts_init)
+            for exchange in self._exchanges.values():
+                exchange.process(data.ts_init)
             self.iteration += 1
             data = self._next()
+        # ---------------------------------------------------------------------#
+        # Process remaining messages
+        for exchange in self._exchanges.values():
+            exchange.process(self._test_clock.timestamp_ns())
         # ---------------------------------------------------------------------#
 
         if not streaming:
             self.trader.stop()
+            # Process remaining messages
+            for exchange in self._exchanges.values():
+                exchange.process(self._test_clock.timestamp_ns())
             self._post_run(
                 run_started=run_started,
                 run_finished=self._clock.utc_now(),
@@ -773,11 +795,6 @@ cdef class BacktestEngine:
             self._test_clock.set_time(event_handler.event.ts_event)
             event_handler.handle()
         self._test_clock.set_time(now_ns)
-
-    cdef void _process_modules(self, int64_t now_ns) except *:
-        cdef SimulatedExchange exchange
-        for exchange in self._exchanges.values():
-            exchange.process_modules(now_ns)
 
     def _pre_run(
         self,
@@ -880,6 +897,7 @@ cdef class BacktestEngine:
             self._log.info("-----------------------------------------------------------------")
             for statistic in self.analyzer.get_performance_stats_returns_formatted():
                 self._log.info(statistic)
+            self._log.info("-----------------------------------------------------------------")
 
     def _add_data_client_if_not_exists(self, ClientId client_id) -> None:
         if client_id not in self._data_engine.registered_clients():