+++ nautilus_trader/common/actor.pyx
@@ -46,7 +46,7 @@ from nautilus_trader.data.messages cimport DataRequest
 from nautilus_trader.data.messages cimport DataResponse
 from nautilus_trader.data.messages cimport Subscribe
 from nautilus_trader.data.messages cimport Unsubscribe
-from nautilus_trader.model.c_enums.book_level cimport BookLevel
+from nautilus_trader.model.c_enums.book_type cimport BookType
 from nautilus_trader.model.data.bar cimport Bar
 from nautilus_trader.model.data.bar cimport BarType
 from nautilus_trader.model.data.base cimport DataType
@@ -630,7 +630,7 @@ cdef class Actor(Component):
     cpdef void subscribe_order_book_deltas(
         self,
         InstrumentId instrument_id,
-        BookLevel level=BookLevel.L2,
+        BookType book_type=BookType.L2_MBP,
         dict kwargs=None,
     ) except *:
         """
@@ -641,8 +641,8 @@ cdef class Actor(Component):
         ----------
         instrument_id : InstrumentId
             The order book instrument ID to subscribe to.
-        level : BookLevel {``L1``, ``L2``, ``L3``}
-            The order book level.
+        book_type : BookType {``L1_TBBO``, ``L2_MBP``, ``L3_MBO``}
+            The order book type.
         kwargs : dict, optional
             The keyword arguments for exchange specific parameters.
 
@@ -660,7 +660,7 @@ cdef class Actor(Component):
             client_id=ClientId(instrument_id.venue.value),
             data_type=DataType(OrderBookData, metadata={
                 "instrument_id": instrument_id,
-                "level": level,
+                "book_type": book_type,
                 "kwargs": kwargs,
             }),
             command_id=self._uuid_factory.generate(),
@@ -672,7 +672,7 @@ cdef class Actor(Component):
     cpdef void subscribe_order_book_snapshots(
         self,
         InstrumentId instrument_id,
-        BookLevel level=BookLevel.L2,
+        BookType book_type=BookType.L2_MBP,
         int depth=0,
         int interval_ms=1000,
         dict kwargs=None,
@@ -688,8 +688,8 @@ cdef class Actor(Component):
         ----------
         instrument_id : InstrumentId
             The order book instrument ID to subscribe to.
-        level : BookLevel {``L1``, ``L2``, ``L3``}
-            The order book level.
+        book_type : BookType {``L1_TBBO``, ``L2_MBP``, ``L3_MBO``}
+            The order book type.
         depth : int, optional
             The maximum depth for the order book. A depth of 0 is maximum depth.
         interval_ms : int
@@ -721,7 +721,7 @@ cdef class Actor(Component):
             client_id=ClientId(instrument_id.venue.value),
             data_type=DataType(OrderBook, metadata={
                 "instrument_id": instrument_id,
-                "level": level,
+                "book_type": book_type,
                 "depth": depth,
                 "interval_ms": interval_ms,
                 "kwargs": kwargs,