+++ tests/integration_tests/adapters/betfair/test_betfair_backtest.py
@@ -25,7 +25,7 @@ from nautilus_trader.model.data.tick import TradeTick
 from nautilus_trader.model.data.venue import InstrumentClosePrice
 from nautilus_trader.model.data.venue import InstrumentStatusUpdate
 from nautilus_trader.model.enums import AccountType
-from nautilus_trader.model.enums import BookLevel
+from nautilus_trader.model.enums import BookType
 from nautilus_trader.model.enums import OMSType
 from nautilus_trader.model.enums import VenueType
 from nautilus_trader.model.identifiers import ClientId
@@ -78,7 +78,7 @@ def create_engine(instruments, data):
         account_type=AccountType.CASH,
         base_currency=GBP,
         starting_balances=[Money(100_000, GBP)],
-        order_book_level=BookLevel.L2,
+        book_type=BookType.L2_MBP,
     )
     return engine